change endpoint from int to float64

+++ mymath.go
@@ -4,6 +4,6 @@ package mymath
 import "math"
 
 // Sqrt - функция для вычисления квадратного корня из числа
-func Sqrt(x float64) int {
+func Sqrt(x float64) float64 {
     return math.Sqrt(x)
 }
